add auth hook, centralize styling
