+++ Mocktologist/App.js
@@ -1,9 +1,14 @@
 import 'react-native-gesture-handler';
 import React from 'react';
 import Navigation from './Navigation';
+import { AuthProvider } from './hooks/useAuth';
 
 const App = () => {
-  return <Navigation />;
+  return (
+    <AuthProvider>
+      <Navigation />
+    </AuthProvider>
+  )
 };
 
 export default App;
