+++ Mocktologist/pages/dash.js
@@ -3,7 +3,7 @@ import { View, Text } from "react-native";
 export default function Dash() {
     return (
         <View >
-            <Text> Dash Page </Text>
+            <Text> Dashboard </Text>
         </View>
     );
 }
